land section completed and fix some errors

+++ src/app/api/attendance/[id]/delete/[workerId]/route.ts
@@ -4,26 +4,26 @@ import ErrorHandler from "@/utils/errorHandler";
 import { wrapAsync } from "@/utils/wrapAsync";
 import { getServerSession, User } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
-import { authOptions } from "../../auth/[...nextauth]/options";
+import { authOptions } from "../../../../auth/[...nextauth]/options";
 
-export const DELETE = wrapAsync(async (req: NextRequest) => {
+export const DELETE = wrapAsync(async (req: NextRequest,{params}:{params : {id : string, workerId : string}}) => {
   await dbConnect();
 
   const session = await getServerSession(authOptions);
   const user: User = session?.user as User;
 
-  if(!session || !session.user){
-      throw new ErrorHandler("Not Authenticated",400)
-    }
+  if (!session || !session.user) {
+    throw new ErrorHandler("Not Authenticated", 400);
+  }
 
-  const { id } = await req.json();
+  const { id, workerId } = await params
 
-  if (!id) {
-    throw new ErrorHandler("Provide the ID", 400);
+  if (!id || !workerId) {
+    throw new ErrorHandler("Provide the ID AND workerID", 400);
   }
 
   const attendace = await AttendanceModel.findOneAndDelete({
-    $and: [{ _id: id }, {}],
+    $or: [{ _id: id }, { worker: workerId }],
   });
   if (!attendace) {
     throw new ErrorHandler("Invalid ID", 400);
